Hibernate oppgradering brekker paging brukt i konsistenavstemming. Endrer til Long
Hibernate oppgradering brekker paging brukt i konsistenavstemming. Endrer til Long

java.lang.ClassCastException: class java.lang.Long cannot be cast to class java.math.BigInteger (java.lang.Long and java.math.BigInteger are in module java.base of loader 'bootstrap')

+++ src/main/kotlin/no/nav/familie/ks/sak/kjerne/behandling/BehandlingService.kt
@@ -41,7 +41,6 @@ import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import java.math.BigInteger
 
 @Service
 class BehandlingService(
@@ -191,7 +190,7 @@ class BehandlingService(
         oppdaterBehandling(behandling)
     }
 
-    fun hentSisteIverksatteBehandlingerFraLøpendeFagsaker(page: Pageable): Page<BigInteger> =
+    fun hentSisteIverksatteBehandlingerFraLøpendeFagsaker(page: Pageable): Page<Long> =
         behandlingRepository.finnSisteIverksatteBehandlingFraLøpendeFagsaker(page)
 
     fun hentAktivtFødselsnummerForBehandlinger(behandlingIder: List<Long>): Map<Long, String> =
